+++ neo/Core/BlockBase.py
@@ -1,4 +1,3 @@
-import ctypes
 from .Mixins import VerifiableMixin
 from neocore.Cryptography.Helper import bin_dbl_sha256
 from neocore.Cryptography.Crypto import Crypto
@@ -7,6 +6,7 @@ from neo.Core.Helper import Helper
 from neo.Blockchain import GetBlockchain, GetGenesis
 from neo.Core.Witness import Witness
 from neocore.UInt256 import UInt256
+from neo.Core.Size import Size as s
 
 
 class BlockBase(VerifiableMixin):
@@ -96,12 +96,10 @@ class BlockBase(VerifiableMixin):
         Returns:
             int: size.
         """
-        uintsize = ctypes.sizeof(ctypes.c_uint)
-        ulongsize = ctypes.sizeof(ctypes.c_ulong)
         scriptsize = 0
         if self.Script is not None:
             scriptsize = self.Script.Size()
-        return uintsize + self.PrevHash.Size + self.MerkleRoot.Size + uintsize + uintsize + ulongsize + 160 + 1 + scriptsize
+        return s.uint32 + s.uint256 + s.uint256 + s.uint32 + s.uint32 + s.uint64 + s.uint160 + 1 + scriptsize
 
     def IndexBytes(self):
         """
@@ -212,9 +210,7 @@ class BlockBase(VerifiableMixin):
         """
         json = {}
         json["hash"] = self.Hash.To0xString()
-
-        # todo: this size isn't calculating correctly currently, but it should be included in the JSON
-        # json["size"] = self.Size()
+        json["size"] = self.Size()
         json["version"] = self.Version
         json["previousblockhash"] = self.PrevHash.To0xString()
         json["merkleroot"] = self.MerkleRoot.To0xString()